Collect pybind11 wrapping code of the viewer to the single file
